fixed some cpp98 uncompatibilites

+++ cpp06/ex01/Serializer.cpp
@@ -8,12 +8,12 @@ Serializer &Serializer::operator=(const Serializer &right) {(void)right;return(*
 
 Serializer::~Serializer(void) {}
 
-uintptr_t	Serializer::serialize(Data*ptr)
+unsigned long	Serializer::serialize(Data*ptr)
 {
-	return (reinterpret_cast<uintptr_t>(ptr));
+	return (reinterpret_cast<unsigned long>(ptr));
 }
 
-Data*		Serializer::deserialize(uintptr_t raw)
+Data*		Serializer::deserialize(unsigned long raw)
 {
 	return (reinterpret_cast<Data*>(raw));
 }
